+++ src/terminal/InputGenerator.h
@@ -311,7 +311,7 @@ class InputGenerator
     std::string_view peek() const noexcept
     {
         return std::string_view(pendingSequence_.data() + consumedBytes_,
-                                pendingSequence_.size() - consumedBytes_);
+                                pendingSequence_.size() - (size_t) consumedBytes_);
     }
 
     void consume(int n)
